Heavy update on Swix
The initial purpose of this was to update swix to be scrubbed down and ready to
post on Kongregate.  In doing so several small incomplete details are getting
wrapped up.

Switched to a more legible font
Added music and sound effects
Added a main menu
Added favicons to get rid of the console warnings

Further actions needed

 - Add an about/credits section on the main menu, crediting the appropriate
   creators for music, sound effects, and fonts
 - Add a congratulatory screen on completion
 - Find a better way to give instructions.  Maybe a "how to play" option on the
   front page, or better instructional technique in the gameplay.

Other "maybe" items:
 - Integrate some of Kongregate's api components to store player's best scores
   or other data
 - Add levels
 - Remove jQuery