Created a preferences subscreen
Previously there was only a sound on/off toggle in the menu.  That has been
moved into a submenu with the same flipping style as the about menu.  It now
has separate volume levels for the music and sound effects, as well as a simple
sound on/off switch.  Any other settings will be added there, including
"current level", letting you choose what level to start at.