Quick bug fixes spotted in Kongregate
1) The game was too tall, adding scrollbars.  I was able to fix this by
lowering the "swix" logo at the top and reducing the upper margin above
the game.

2) The very large sprite that is used as a backdrop in the main menu
takes a long time to load.  Added a chunk of code to pre-cache it in the
load sequence.  This will make the load time slightly longer, but the
front menu won't be a mess as a result.