Fixed info prompt legibility
There are two levels in this game that start with a "tip" explaining game flow
and operation.  That tip was previously using a barely legible font in some
platforms, and was generally small and difficult to read.  Fixed that, using a
clear font, making it larger, and making the buttons prettier.

Also added a top margin to the cells when pressed, which adds a lot more of a
tactile feel to the game I think.  Just a CSS one-liner, so if others find it
offputting it can be removed easily.