Full game loop complete
We now have a full playable cycle of launching, watching the ball bounce around
and break bricks, and the bricks descending to the bottom with the game
restarting there.
From here on it's a matter of adding basic features (ball additions, title
screen, etc.
Currently the ball treats the brick like a circle, bouncing off with the same
speed, but folling the unit vector of the middle of the block to the point of
collision.  In the future I should use that as a line of reflection, and mirror
the original motion vector across that.