Changed ball bounce logic.
It's still not perfect.  Previously it would bounce (incorrectly) off of the
blocks like they were circles.  It was calculated wrong anyway, but looked ok.
It now carefully measures out which edges are being hit and bounces off
accordingly.  That's pretty decent except that it is not perfect on corners,
and makes the ball motion too boring, as the angle never changes.

A potential solution to this would be to also take the relative angle of the
vector from box centre to ball centre, and if it's within a reasonable range of
a corner (pi / 4) + (pi * n / 2) I think.. calculate the bounce angle based on
its angle.