Fixed collision with edges
There was an edge case where the launch point could be far enough to the side
that larger balls would get stuck moving vertically because their x-direction
would be inverted on each step.  That's been corrected by first checking to see
whether the velocity is already in the desired direction.