Improved collision
Actually less accurate now, but it has better looking rebound angles off of
corners.

Also added a couple of functions in geometry.js, one that returns a unit vector
for a given vector, and one that uses two given points to define a line across
which a third point is reflected.