Fixed scaling issue with balls
There was a constant being used to calculate the ball radius relative to the
game scale, but it was hard coded, and not being used in checking collision
between the ball and the bricks, causing collision to happen at incorrect
distances from the block.  Fixed that by adding a global constant and using it
in the calculations.