Added bonus blocks
Instead of having blocks that occasionally have double the strength, I now have
bonus blocks.  Those start off with the expected count, and have a texas star
oscillating in front of them.  When you break them, their count is reset, and
they now have three stars oscillating in front.  When that one is broken, the
block is gone, and you currently get 10 * the block's original value in points.
Eventually you will also get a power-up with it as well.