Have a better effect in place for bomb bonus
When a block with the bomb bonus is broken, instead of giving you a weird bonus
bomb ball, you get a pile of balls (16 I think) radiating outward from the
center of the broken block.  Much more satisfying.

The next step will be to indicate the bonus with an icon in the block, rather
than leaving it unknown.