Now have speed increasing bonuses
Had to refactor quite a bit of code to get this to work.  Previously I wrote
this to provide a bonus "object" that would fall upon breaking a bonus block,
but that didn't have the right flow to it.  Bonuses now happen directly from
the block itself.

There are now two working bonuses in place, blocks that explode into extra
balls that aren't retained, and the launch frequency and velocity of the balls
increasing.