Bad code fix
In inaders, I had to add a custom json_enccode function to run on one of the
servers its on, as it's using an old enough version of php that it wasn't an
existing function.  I made the mistake though of not checking whether that
function exist already, causing errors when hosted on other sites.

Fixed