An initial commit adding new Nonogram game
I wanted a Nonogram game, but any I've found have flaws that bug me, so I
thought I'd write my own.

This is the initial commit throwing in everything I've got so far.  It's not a
functioning game yet, but it does generate and render the board properly. User
interaction to follow shortly.