Cleaned up interface and code
Got rid of the main menu, opting to simply have buttons in the game.  Currently
those buttons are at the top of the page with the options "new" which generates
a new puzzle, and "exit" which returns to the previous URL.

Cleaned up the CSS as well, removing a bunch of stuff that was unnecessary.